Move XLNet config and modeling code to xlnet folder.
PiperOrigin-RevId: 280720540
